requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
moviepy==1.0.3
numpy==1.19.1
Pillow==7.2.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1

# Added in manually! Do not remove
gunicorn==19.6.0